test: move classes implementing Jupiter interfaces to their own subpackage
