+++ engine-tests/src/test/java/org/terasology/engine/integrationenvironment/ModuleTestingEnvironmentTest.java
@@ -10,6 +10,7 @@ import com.google.common.util.concurrent.UncheckedTimeoutException;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.terasology.engine.integrationenvironment.jupiter.MTEExtension;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
